+++ Source/Machine.VSTestAdapter/Machine.VSTestAdapter.csproj
@@ -78,17 +78,17 @@
     <Compile Include="MspecTestAdapterExecutor.cs" />
     <Compile Include="MSpecVSRunnerManager.cs" />
     <Compile Include="ISpecificationExecutor.cs" />
-    <Compile Include="ISpecificationDiscoverer.cs" />
+    <Compile Include="Discovery\ISpecificationDiscoverer.cs" />
     <Compile Include="MSpecTestAdapter.cs" />
     <Compile Include="MSpecTestAdapterFactory.cs" />
-    <Compile Include="MSpecTestCase.cs" />
+    <Compile Include="Discovery\MSpecTestCase.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RunStats.cs" />
     <Compile Include="Discovery\Cecil\ScopedAssemblyResolver.cs" />
     <Compile Include="Discovery\Cecil\CecilSpecificationDiscoverer.cs" />
     <Compile Include="SpecificationExecutor.cs" />
     <Compile Include="SpecificationRunListener.cs" />
-    <Compile Include="SpecTestHelper.cs" />
+    <Compile Include="Discovery\SpecTestHelper.cs" />
     <Compile Include="Strings.Designer.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
